Added R script used for the revision
This script performs the analyses showed in Morales et al. (under review). This script focuses only on the response-locked data from the Zoo task.
The script starts by organizing all the data and then performing the analyses. First the age-related analyses are done followed by the reliability analyses. 
The script is more like a notebook rather than a script that can be run from start to finish. Running the script from start to finish would take too long because of the reliability analyses.